+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## NEW FUNCTIONALITY
 
-* `statistics/spatial_statistics`: Added a component to calculate spatial statistics on Xenium-ingested data (PR #46).
+* `feature_annotation/xenium_spatial_statistics`: Added a component to calculate spatial statistics on Xenium-ingested data (PR #46).
 
 # openpipeline_spatial 0.3.0