+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## NEW FUNCTIONALITY
 
+* `feature_annotation/xenium_spatial_statistics`: Added a component to calculate spatial statistics on Xenium-ingested data (PR #46).
+
 * `neighbors/join_graphs`: Combine spatial and expression neighborhood graphs into a single graph (PR #47).
 
 # openpipeline_spatial 0.3.0